requirements.txt
pandas>=2.0.0
openai>=1.0.0
deepl>=1.15.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
python-dotenv>=1.0.0
